--- a/.changeset/great-rice-return.md
+++ b/.changeset/great-rice-return.md
@@ -1,0 +1,5 @@
+---
+'markdown-to-jsx': patch
+---
+
+Further optimize the plain text splitting regex.