--- a/.changeset/grumpy-kids-attack.md
+++ b/.changeset/grumpy-kids-attack.md
@@ -1,0 +1,5 @@
+---
+'markdown-to-jsx': patch
+---
+
+Remove redundant detectors when processing paragraphs.