--- a/.changeset/hungry-bugs-tan.md
+++ b/.changeset/hungry-bugs-tan.md
@@ -1,0 +1,5 @@
+---
+'markdown-to-jsx': patch
+---
+
+Replace some regexes with optimized functions to avoid polynomial time scenarios. Also fixes compatibility issues in some older browsers with the `trimEnd` API.