--- a/.changeset/plenty-dodos-collect.md
+++ b/.changeset/plenty-dodos-collect.md
@@ -1,0 +1,5 @@
+---
+'markdown-to-jsx': patch
+---
+
+Rework inline code syntax handling, handle escaped characters in code blocks correctly so they render without the backslash.