--- a/.prettierignore
+++ b/.prettierignore
@@ -1,0 +1,1 @@
+fixture.md