--- a/package.json
+++ b/package.json
@@ -66,6 +66,7 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-serializer-html": "^7.1.0",
+    "jest-watch-typeahead": "^2.2.2",
     "markdown-it": "^14.0.0",
     "microbundle": "^0.15.1",
     "microtime": "^3.1.1",
@@ -122,6 +123,10 @@
     },
     "snapshotSerializers": [
       "jest-serializer-html"
+    ],
+    "watchPlugins": [
+      "jest-watch-typeahead/filename",
+      "jest-watch-typeahead/testname"
     ]
   },
   "packageManager": "yarn@4.6.0"