… .changeset/nervous-suns-roll.md
… __snapshots__/index.compiler.spec.tsx.snap
… yarn.lock