```
+--------------------------+------------------------+-----------------------+
|                          │ simple markdown string │ large markdown string |
+--------------------------+------------------------+-----------------------+
| markdown-to-jsx (next)   │ 92,269 ops/sec         │ 330 ops/sec           |
+--------------------------+------------------------+-----------------------+
| markdown-to-jsx (latest) │ 90,360 ops/sec         │ 303 ops/sec           |
+--------------------------+------------------------+-----------------------+
```